# Express-js/swagger.yaml
openapi: 3.0.0
info:
  description: Refactored API
  version: "1.0.0"
  title: Project
paths:
...
  agreement/{agreementId}/item/{action}:
    post:
      summary: add / remove a item to agreement
      operationId: agreementItem
      security:
        - api_key: []
      parameters:
        - name: agreementId
          in: path
          description: agreement Id
          required: true
          schema:
            type: integer
        - name: action
          in: path
          description: action (add / remove)
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/Item'
          
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/responses/Success'
...
components:
  schemas:
...
    Item:
      type: object
      properties:
        itemId:
          type: integer
        removeAll:
          type: boolean
        itemType:
          type: string
  requestBodies:
...
    Item:
      content: 
        application/json:
          schema:
            $ref: '#/components/schemas/Item'
  responses:  
    Success:
      description: success
  securitySchemes:
    api_key:
      type: apiKey
      name: token
      in: header